complete overhaul i gotta commit more lmao

+++ app/Http/Controllers/RomhackVersionsController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreRomhackVersionRequest;
+use App\Http\Requests\UpdateRomhackVersionRequest;
 use App\Models\Author;
 use App\Models\Romhack;
 use App\Models\Version;
@@ -16,32 +18,31 @@ class RomhackVersionsController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request, Romhack $hack, ?Version $version = null)
+    public function store(StoreRomhackVersionRequest $request, Romhack $hack)
     {
-        $r = $request->validate(
-            [
-                'version.name' => 'required|string',
-                'version.starcount' => 'required|integer|numeric|gte:0',
-                'version.releasedate' => 'required|date',
-                'version.filename' => 'exclude_if:_method,PUT|required|file|filled|extensions:zip',
-                'version.demo' => 'required|bool',
-                'version.recommened' => 'required|bool',
-                'version.author.name' => 'required|array|min:1',
-                'version.author.name.*' => 'required|string'
-            ]
-        );
+        $r = $request->validated();
         $authorNames = Arr::get(Arr::get(Arr::get($r, 'version'), 'author'), 'name');
-        if ($request->isMethod('POST')) {
-            $path = $request->file('version.filename')->store('patch', 'public');
-            Arr::set($r, 'version.filename', $path);
-        }
+        $path = $request->file('version.filename')->store('patch', 'public');
+        Arr::set($r, 'version.filename', $path);
 
-        if (is_null($version)) {
-            $version = $hack->versions()->create(Arr::get($r, 'version'));
-        } else {
-            $version->update(Arr::get($r, 'version'));
-            $version->authors()->detach();
+        $version = $hack->versions()->create(Arr::get($r, 'version'));
+        foreach ($authorNames as $author) {
+            $a = Author::createOrFirst(['name' => $author]);
+            $version->authors()->attach($a->id);
         }
+        return redirect(route('hack.show', ['hack' => $hack]));
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     */
+    public function update(UpdateRomhackVersionRequest $request, Romhack $hack, Version $version)
+    {
+        $r = $request->validated();
+        $authorNames = Arr::get(Arr::get(Arr::get($r, 'version'), 'author'), 'name');
+        $version->update(Arr::get($r, 'version'));
+        $version->authors()->detach();
+
         foreach ($authorNames as $author) {
             $a = Author::createOrFirst(['name' => $author]);
             $version->authors()->attach($a->id);
@@ -52,28 +53,38 @@ class RomhackVersionsController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function manage(Romhack $hack, ?Version $version)
+    public function create(Romhack $hack)
     {
         $authors = Author::all()->sortBy('name');
-        return view('hacks.versions.manage', ['version' => $version, 'authors' => $authors]);
+        return view('hacks.versions.create', ['hack' => $hack, 'authors' => $authors]);
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     */
+    public function edit(Romhack $hack, Version $version)
+    {
+        $authors = Author::all()->sortBy('name');
+        return view('hacks.versions.edit', ['version' => $version, 'authors' => $authors]);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function delete(Romhack $hack, Version $version)
+    public function destroy(Request $request, Romhack $hack, Version $version)
     {
-        return view('hacks.versions.delete', ['hack' => $hack, 'version' => $version]);
-    }
+        if ($request->user()->cannot('delete', $version)) {
+            abort(403);
+        }
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Romhack $hack, Version $version)
-    {
         $version->delete();
         Storage::disk('public')->delete($version->filename);
-        dd($version);
+        if ($hack->versions()->count() > 0) {
+            return redirect(route('hack.show', ['hack' => $hack]));
+        } else {
+            $hack->delete();
+            return redirect(route('hack.index'));
+        }
     }
 
     /**
